pyproject.toml
[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic_ai"
version = "0.1.0"
description = "Retrieval-Augmented Generation with agent workflows"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name="Your Name"}]
dependencies = [
    "faiss-cpu>=1.8.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.7.0",
    "transformers>=4.44.0",
    "accelerate>=0.33.0",
    "peft>=0.11.0",
    "datasets>=2.20.0",
    "torch>=2.2.0",
    "fastapi>=0.114.0",
    "uvicorn>=0.30.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
    "openai>=1.43.0",
    "anthropic>=0.34.1",
    "tqdm>=4.66.0",
    "pdfminer.six>=20240706",
    "beautifulsoup4>=4.12.3",
    "streamlit>=1.38.0",
]

[project.scripts]
agentic-ai = "agentic_ai.cli:app"
